pyproject.toml
[project]
name = "vol-surface-opt-trans"
version = "0.1.0"
description = "Reproducible pipeline for SPX options vol surfaces and distribution distances"
requires-python = ">=3.11"
dependencies = [
    "polars>=0.20",
    "numpy>=1.24",
    "scipy>=1.11",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
    "matplotlib>=3.7",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
garch = ["arch>=6.0"]
dev = ["pytest>=7.0", "ruff>=0.1"]

[project.scripts]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "pipeline"]
